Rename Contact.tsx to Contact.tsx
